refactor(service): Service 인터페이스 구현 및 컨트롤러와 느슨한 의존으로 개선
