+++ src/main/java/wooteco/subway/ui/StationController.java
@@ -1,22 +1,18 @@
-package wooteco.subway.controller;
+package wooteco.subway.ui;
 
 import java.net.URI;
 import java.util.List;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import wooteco.subway.repository.exception.DuplicateStationNameException;
 import wooteco.subway.service.StationService;
-import wooteco.subway.service.dto.station.StationRequest;
-import wooteco.subway.service.dto.station.StationResponse;
+import wooteco.subway.ui.dto.request.StationRequest;
+import wooteco.subway.ui.dto.response.StationResponse;
 
 @RestController
 @RequestMapping("/stations")
@@ -30,25 +26,20 @@ public class StationController {
 
     @PostMapping
     public ResponseEntity<StationResponse> createStation(@RequestBody StationRequest stationRequest) {
-        StationResponse stationResponse = stationService.create(stationRequest.getName());
-        URI redirectUri = URI.create("/stations/" + stationResponse.getId());
+        final StationResponse stationResponse = stationService.create(stationRequest.getName());
+        final URI redirectUri = URI.create("/stations/" + stationResponse.getId());
         return ResponseEntity.created(redirectUri).body(stationResponse);
     }
 
     @GetMapping
     public ResponseEntity<List<StationResponse>> showStations() {
-        List<StationResponse> stationResponses = stationService.findAll();
+        final List<StationResponse> stationResponses = stationService.findAll();
         return ResponseEntity.ok(stationResponses);
     }
 
-    @DeleteMapping("/{stationId}")
-    public ResponseEntity<Void> deleteStation(@PathVariable Long stationId) {
-        stationService.remove(stationId);
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteStation(@PathVariable Long id) {
+        stationService.deleteById(id);
         return ResponseEntity.noContent().build();
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<String> handle(DuplicateStationNameException exception) {
-        return ResponseEntity.badRequest().body(exception.getMessage());
     }
 }
